storage: move store version checks into fs pkg
Move functions to perform pebble compatibility checks into the fs pkg.
